Refactor Options & Option component
