+++ src/DropdownMenu.js
@@ -1,5 +1,5 @@
 import React, { Component, isValidElement, cloneElement } from 'react';
-import OptionWrapper from './OptionWrapper';
+import Options from './Options';
 
 const renderComponent = (Component, props) => {
   if (isValidElement(Component)) {
@@ -12,60 +12,15 @@ const renderComponent = (Component, props) => {
 };
 
 export default class DropdownMenu extends Component {
-  componentWillReceiveProps(nextProps) {
-    this.scrollTo(nextProps.highlightedIndex);
-  }
-
-  componentDidMount() {
-    this.optionsListOffsetHeight = this.optionsList.offsetHeight;
-    this.stashOptionOffsetHeight();
-    this.scrollTo(this.props.highlightedIndex);
-  }
-
-  componentDidUpdate() {
-    if (!this.optionOffsetHeight) {
-      this.stashOptionOffsetHeight();
-    }
-
-    if (!this.optionsListOffsetHeight) {
-      this.optionsListOffsetHeight = this.optionsList.offsetHeight;
-    }
-  }
-
-  stashOptionOffsetHeight() {
-    let option = document.querySelector('.PowerSelect__Option');
-    this.optionOffsetHeight = (option && option.offsetHeight) || 0;
-  }
-
-  scrollTo(newHighlightedIndex) {
-    if (newHighlightedIndex !== this.props.highlightedIndex) {
-      let optionOffsetHeight = this.optionOffsetHeight;
-      let delta =
-        optionOffsetHeight * newHighlightedIndex +
-        optionOffsetHeight -
-        this.optionsListOffsetHeight;
-
-      if (delta > 0) {
-        this.optionsList.scrollTop = delta;
-      } else {
-        this.optionsList.scrollTop = 0;
-      }
-    }
-  }
-
   render() {
     let {
-      options,
-      onOptionClick,
-      handleKeyDown,
-      selected,
-      optionLabelPath,
-      optionComponent,
       select,
-      minWidth,
+      handleKeyDown,
       highlightedIndex,
+      minWidth,
       beforeOptionsComponent,
       afterOptionsComponent,
+      ...otherProps
     } = this.props;
 
     return (
@@ -78,24 +33,13 @@ export default class DropdownMenu extends Component {
       >
         {beforeOptionsComponent &&
           renderComponent(beforeOptionsComponent, { select })}
-        <div
-          className="PowerSelect__Options"
-          ref={optionsList => (this.optionsList = optionsList)}
-        >
-          {options.map((option, idx) =>
-            <OptionWrapper
-              key={idx}
-              option={option}
-              select={select}
-              optionLabelPath={optionLabelPath}
-              optionComponent={optionComponent}
-              isHighlighted={highlightedIndex === idx}
-              onOptionClick={() => {
-                onOptionClick(idx, option);
-              }}
-            />
-          )}
-        </div>
+
+        <Options
+          select={select}
+          highlightedIndex={highlightedIndex}
+          {...otherProps}
+        />
+
         {afterOptionsComponent &&
           renderComponent(afterOptionsComponent, { select })}
       </div>
