+++ src/PowerSelect/index.js
@@ -14,6 +14,7 @@ export default class PowerSelect extends Component {
   }
 }
 
+PowerSelect.displayName = 'PowerSelect';
 PowerSelect.defaultProps = {
   searchEnabled: true,
   beforeOptionsComponent: SearchInput,
