+++ src/TypeAhead/index.js
@@ -20,3 +20,5 @@ export default class TypeAhead extends Component {
     );
   }
 }
+
+TypeAhead.displayName = 'TypeAhead';
